+++ src/loader/parsers/compressed_textures/parsePKM.js
@@ -1,0 +1,5 @@
+
+// parse and return the compressed texture
+export function parsePKM(/*data, formats*/) {
+    throw new Error("unsupported format");
+}