+++ src/loader/parsers/compressed_textures/parsePVR.js
@@ -1,0 +1,5 @@
+
+// parse and return the compressed texture
+export function parsePVR(/*data, formats*/) {
+    throw new Error("unsupported format");
+}